[BREAKING] Add version.py and package.json support
This is breaking change, the .crom.toml file will need to be updated to
use this version.

The .crom.toml file moved `types` to being a root level object. The
README.md was updated to match this new change.

Also PR's and build will now have to pass `cargo fmt` and `cargo clippy`
before they can be merged / built.

Moving the build to be fully in Dockerfiles helping removing some code
from the azure-pipeline.yml

Also formatted the codebase with `cargo fmt` (most fo the changes).