Update the Command Line to be more usable
The changes remove the assumptions that crom makes and instead makes,
the user make the assertions about which type of version.

This was done based on the use of Crom. Most of the time, you have to
override the CLI and force specific formats. Instead of having to
override values, the CLI will now require the user to specify if the
version should be:
- latest
- next-release
- pre-release
- custom

Different sub commands have different options.

The pre-release option will use a short git hash so that when you build
with branches you can have multiple artifacts.